RATIS-455. Move exception classes to a org.apache.ratis.protocol.exceptions
