+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/AlreadyClosedException.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -15,21 +15,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.ratis.protocol.exceptions;
 
-package org.apache.ratis.protocol;
+import org.apache.ratis.protocol.RaftException;
 
-import java.io.IOException;
-
-/** Thrown for checksum errors. */
-public class ChecksumException extends IOException {
-  private static final long serialVersionUID = 1L;
-  private long pos;
-  public ChecksumException(String description, long pos) {
-    super(description);
-    this.pos = pos;
+/**
+ * The corresponding object is already closed.
+ */
+public class AlreadyClosedException extends RaftException {
+  public AlreadyClosedException(String message) {
+    super(message);
   }
 
-  public long getPos() {
-    return pos;
+  public AlreadyClosedException(String message, Throwable t) {
+    super(message, t);
   }
 }
