+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/LeaderNotReadyException.java
@@ -15,21 +15,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.ratis.protocol;
+package org.apache.ratis.protocol.exceptions;
 
-import java.io.IOException;
+import org.apache.ratis.protocol.RaftGroupMemberId;
+import org.apache.ratis.protocol.exceptions.ServerNotReadyException;
 
 /**
- * Timeout has occurred for a blocking I/O.
+ * This exception is sent from the server to a client. The server has just
+ * become the current leader, but has not committed its first place-holder
+ * log entry yet. Thus the leader cannot accept any new client requests since
+ * it cannot determine whether a request is just a retry.
  */
-public class TimeoutIOException extends IOException {
-  static final long serialVersionUID = 1L;
+public class LeaderNotReadyException extends ServerNotReadyException {
+  private final RaftGroupMemberId serverId;
 
-  public TimeoutIOException(String message) {
-    super(message);
+  public LeaderNotReadyException(RaftGroupMemberId id) {
+    super(id + " is in LEADER state but not ready yet.");
+    this.serverId = id;
   }
 
-  public TimeoutIOException(String message, Throwable throwable) {
-    super(message, throwable);
+  public RaftGroupMemberId getServerId() {
+    return serverId;
   }
 }
