+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/NotLeaderException.java
@@ -15,8 +15,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.ratis.protocol;
+package org.apache.ratis.protocol.exceptions;
 
+import org.apache.ratis.protocol.RaftException;
+import org.apache.ratis.protocol.RaftGroupMemberId;
+import org.apache.ratis.protocol.RaftPeer;
 import org.apache.ratis.util.Preconditions;
 
 import java.util.Collection;
