+++ ratis-common/src/main/java/org/apache/ratis/protocol/exceptions/ServerNotReadyException.java
@@ -15,10 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.ratis.protocol;
+package org.apache.ratis.protocol.exceptions;
 
-public class ReconfigurationInProgressException extends RaftException {
-  public ReconfigurationInProgressException(String message) {
+import org.apache.ratis.protocol.RaftException;
+
+/**
+ * The server is not ready yet.
+ */
+public class ServerNotReadyException extends RaftException {
+  public ServerNotReadyException(String message) {
     super(message);
   }
 }
